[story 5] [henry] refactor: add ParkingAble interface and let ParkingBoy interface extends ParkingBoy

+++ src/main/java/com/oocl/cultivation/SuperSmartParkingBoy.java
@@ -3,15 +3,7 @@ package com.oocl.cultivation;
 import java.util.List;
 import java.util.Optional;
 
-/**
- * @ProjectName: parking-lot
- * @Package: com.oocl.cultivation
- * @ClassName: SuperSmartParkingBoy
- * @Author: carrymaniac
- * @Description:
- * @Date: 2020/7/26 11:27 下午
- * @Version:
- */
+
 public class SuperSmartParkingBoy implements ParkingBoy{
     List<ParkingLot> parkingLots;
     public SuperSmartParkingBoy(List<ParkingLot> parkingLots) {
@@ -22,7 +14,8 @@ public class SuperSmartParkingBoy implements ParkingBoy{
     public ParkResult park(Car car) {
         Optional<ParkingLot> first = parkingLots.stream()
                 .filter(parkingLot -> parkingLot.getRemainingPosition() > 0)
-                .max((a, b) -> a.getRemainingPosition()/a.getCapacity() >= b.getRemainingPosition()/b.getCapacity() ? 1 : -1);
+                //todo
+                .max((a, b) -> (double)a.getRemainingPosition()/(double)a.getCapacity() >= (double)b.getRemainingPosition()/(double)b.getCapacity() ? 1 : -1);
         if (first.isPresent()) {
             return first.get().park(car);
         } else {
